[Post-Australis RUN FOR YOUR GOD DAMNED LIFE!] Manual merge of m-r29

+++ testing/mozbase/mozprocess/mozprocess/processhandler.py
@@ -2,8 +2,6 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
-import logging
-import mozinfo
 import os
 import select
 import signal
@@ -19,7 +17,11 @@ __all__ = ['ProcessHandlerMixin', 'ProcessHandler']
 # Set the MOZPROCESS_DEBUG environment variable to 1 to see some debugging output
 MOZPROCESS_DEBUG = os.getenv("MOZPROCESS_DEBUG")
 
-if mozinfo.isWin:
+# We dont use mozinfo because it is expensive to import, see bug 933558.
+isWin = os.name == "nt"
+isPosix = os.name == "posix" # includes MacOS X
+
+if isWin:
     import ctypes, ctypes.wintypes, msvcrt
     from ctypes import sizeof, addressof, c_ulong, byref, POINTER, WinError, c_longlong
     import winprocess
@@ -28,7 +30,25 @@ if mozinfo.isWin:
     JOBOBJECT_BASIC_LIMIT_INFORMATION, JOBOBJECT_EXTENDED_LIMIT_INFORMATION, IO_COUNTERS
 
 class ProcessHandlerMixin(object):
-    """Class which represents a process to be executed."""
+    """
+    A class for launching and manipulating local processes.
+
+    :param cmd: command to run. May be a string or a list. If specified as a list, the first element will be interpreted as the command, and all additional elements will be interpreted as arguments to that command.
+    :param args: list of arguments to pass to the command (defaults to None). Must not be set when `cmd` is specified as a list.
+    :param cwd: working directory for command (defaults to None).
+    :param env: is the environment to use for the process (defaults to os.environ).
+    :param ignore_children: causes system to ignore child processes when True, defaults to False (which tracks child processes).
+    :param kill_on_timeout: when True, the process will be killed when a timeout is reached. When False, the caller is responsible for killing the process. Failure to do so could cause a call to wait() to hang indefinitely. (Defaults to True.)
+    :param processOutputLine: function to be called for each line of output produced by the process (defaults to None).
+    :param onTimeout: function to be called when the process times out.
+    :param onFinish: function to be called when the process terminates normally without timing out.
+    :param kwargs: additional keyword args to pass directly into Popen.
+
+    NOTE: Child processes will be tracked by default.  If for any reason
+    we are unable to track child processes and ignore_children is set to False,
+    then we will fall back to only tracking the root process.  The fallback
+    will be logged.
+    """
 
     class Process(subprocess.Popen):
         """
@@ -59,7 +79,7 @@ class ProcessHandlerMixin(object):
             # Parameter for whether or not we should attempt to track child processes
             self._ignore_children = ignore_children
 
-            if not self._ignore_children and not mozinfo.isWin:
+            if not self._ignore_children and not isWin:
                 # Set the process group id for linux systems
                 # Sets process group id to the pid of the parent process
                 # NOTE: This prevents you from using preexec_fn and managing
@@ -79,7 +99,7 @@ class ProcessHandlerMixin(object):
                 raise
 
         def __del__(self, _maxint=sys.maxint):
-            if mozinfo.isWin:
+            if isWin:
                 if self._handle:
                     if hasattr(self, '_internal_poll'):
                         self._internal_poll(_deadstate=_maxint)
@@ -90,9 +110,9 @@ class ProcessHandlerMixin(object):
             else:
                 subprocess.Popen.__del__(self)
 
-        def kill(self):
+        def kill(self, sig=None):
             self.returncode = 0
-            if mozinfo.isWin:
+            if isWin:
                 if not self._ignore_children and self._handle and self._job:
                     winprocess.TerminateJobObject(self._job, winprocess.ERROR_CONTROL_C_EXIT)
                     self.returncode = winprocess.GetExitCodeProcess(self._handle)
@@ -106,20 +126,18 @@ class ProcessHandlerMixin(object):
                     self._cleanup()
                     if err is not None:
                         raise OSError(err)
-                else:
-                    pass
             else:
+                sig = sig or signal.SIGKILL
                 if not self._ignore_children:
                     try:
-                        os.killpg(self.pid, signal.SIGKILL)
+                        os.killpg(self.pid, sig)
                     except BaseException, e:
                         if getattr(e, "errno", None) != 3:
                             # Error 3 is "no such process", which is ok
                             print >> sys.stdout, "Could not kill process, could not find pid: %s, assuming it's already dead" % self.pid
                 else:
-                    os.kill(self.pid, signal.SIGKILL)
-                if self.returncode is None:
-                    self.returncode = subprocess.Popen._internal_poll(self)
+                    os.kill(self.pid, sig)
+                self.returncode = -sig
 
             self._cleanup()
             return self.returncode
@@ -137,7 +155,7 @@ class ProcessHandlerMixin(object):
 
         """ Private Members of Process class """
 
-        if mozinfo.isWin:
+        if isWin:
             # Redefine the execute child so that we can track process groups
             def _execute_child(self, args, executable, preexec_fn, close_fds,
                                cwd, env, universal_newlines, startupinfo,
@@ -397,7 +415,7 @@ falling back to not using job objects for managing child processes"""
                         threadalive = self._procmgrthread.is_alive()
                     else:
                         threadalive = self._procmgrthread.isAlive()
-                if self._job and threadalive: 
+                if self._job and threadalive:
                     # Then we are managing with IO Completion Ports
                     # wait on a signal so we know when we have seen the last
                     # process come through.
@@ -492,7 +510,7 @@ falling back to not using job objects for managing child processes"""
                 else:
                     self._handle = None
 
-        elif mozinfo.isMac or mozinfo.isUnix:
+        elif isPosix:
 
             def _wait(self):
                 """ Haven't found any reason to differentiate between these platforms
@@ -503,8 +521,20 @@ falling back to not using job objects for managing child processes"""
 
                 if not self._ignore_children:
                     try:
-                        # os.waitpid returns a (pid, status) tuple
-                        return os.waitpid(self.pid, 0)[1]
+                        # os.waitpid return value:
+                        # > [...] a tuple containing its pid and exit status
+                        # > indication: a 16-bit number, whose low byte is the
+                        # > signal number that killed the process, and whose
+                        # > high byte is the exit status (if the signal number
+                        # > is zero)
+                        # - http://docs.python.org/2/library/os.html#os.wait
+                        status = os.waitpid(self.pid, 0)[1]
+
+                        # For consistency, format status the same as subprocess'
+                        # returncode attribute
+                        if status > 255:
+                            return status >> 8
+                        return -status
                     except OSError, e:
                         if getattr(e, "errno", None) != 10:
                             # Error 10 is "no child process", which could indicate normal
@@ -538,33 +568,20 @@ falling back to not using job objects for managing child processes"""
                  cwd=None,
                  env=None,
                  ignore_children = False,
+                 kill_on_timeout = True,
                  processOutputLine=(),
                  onTimeout=(),
                  onFinish=(),
                  **kwargs):
-        """
-        cmd = Command to run
-        args = array of arguments (defaults to None)
-        cwd = working directory for cmd (defaults to None)
-        env = environment to use for the process (defaults to os.environ)
-        ignore_children = when True, causes system to ignore child processes,
-        defaults to False (which tracks child processes)
-        processOutputLine = handlers to process the output line
-        onTimeout = handlers for timeout event
-        kwargs = keyword args to pass directly into Popen
-
-        NOTE: Child processes will be tracked by default.  If for any reason
-        we are unable to track child processes and ignore_children is set to False,
-        then we will fall back to only tracking the root process.  The fallback
-        will be logged.
-        """
         self.cmd = cmd
         self.args = args
         self.cwd = cwd
         self.didTimeout = False
         self._ignore_children = ignore_children
+        self._kill_on_timeout = kill_on_timeout
         self.keywordargs = kwargs
         self.outThread = None
+        self.read_buffer = ''
 
         if env is None:
             env = os.environ.copy()
@@ -577,11 +594,12 @@ falling back to not using job objects for managing child processes"""
 
         # It is common for people to pass in the entire array with the cmd and
         # the args together since this is how Popen uses it.  Allow for that.
-        if not isinstance(self.cmd, list):
-            self.cmd = [self.cmd]
-
-        if self.args:
-            self.cmd = self.cmd + self.args
+        if isinstance(self.cmd, list):
+            if self.args != None:
+                raise TypeError("cmd and args must not both be lists")
+            (self.cmd, self.args) = (self.cmd[0], self.cmd[1:])
+        elif self.args is None:
+            self.args = []
 
     @property
     def timedOut(self):
@@ -590,7 +608,7 @@ falling back to not using job objects for managing child processes"""
 
     @property
     def commandline(self):
-        """the string value of the command line"""
+        """the string value of the command line (command + args)"""
         return subprocess.list2cmdline([self.cmd] + self.args)
 
     def run(self, timeout=None, outputTimeout=None):
@@ -598,7 +616,8 @@ falling back to not using job objects for managing child processes"""
         Starts the process.
 
         If timeout is not None, the process will be allowed to continue for
-        that number of seconds before being killed.
+        that number of seconds before being killed. If the process is killed
+        due to a timeout, the onTimeout handler will be called.
 
         If outputTimeout is not None, the process will be allowed to continue
         for that number of seconds without producing any output before
@@ -618,23 +637,27 @@ falling back to not using job objects for managing child processes"""
         args.update(self.keywordargs)
 
         # launch the process
-        self.proc = self.Process(self.cmd, **args)
+        self.proc = self.Process([self.cmd] + self.args, **args)
 
         self.processOutput(timeout=timeout, outputTimeout=outputTimeout)
 
-    def kill(self):
+    def kill(self, sig=None):
         """
-          Kills the managed process and if you created the process with
-          'ignore_children=False' (the default) then it will also
-          also kill all child processes spawned by it.
-          If you specified 'ignore_children=True' when creating the process,
-          only the root process will be killed.
+        Kills the managed process.
 
-          Note that this does not manage any state, save any output etc,
-          it immediately kills the process.
+        If you created the process with 'ignore_children=False' (the
+        default) then it will also also kill all child processes spawned by
+        it. If you specified 'ignore_children=True' when creating the
+        process, only the root process will be killed.
+
+        Note that this does not manage any state, save any output etc,
+        it immediately kills the process.
+
+        :param sig: Signal used to kill the process, defaults to SIGKILL
+                    (has no effect on Windows)
         """
         try:
-            return self.proc.kill()
+            return self.proc.kill(sig=sig)
         except AttributeError:
             # Try to print a relevant error message.
             if not self.proc:
@@ -644,21 +667,21 @@ falling back to not using job objects for managing child processes"""
 
     def readWithTimeout(self, f, timeout):
         """
-          Try to read a line of output from the file object |f|.
-          |f| must be a  pipe, like the |stdout| member of a subprocess.Popen
-          object created with stdout=PIPE. If no output
-          is received within |timeout| seconds, return a blank line.
-          Returns a tuple (line, did_timeout), where |did_timeout| is True
-          if the read timed out, and False otherwise.
+        Try to read a line of output from the file object *f*.
 
-          Calls a private member because this is a different function based on
-          the OS
+        *f* must be a  pipe, like the *stdout* member of a subprocess.Popen
+        object created with stdout=PIPE. If no output
+        is received within *timeout* seconds, return a blank line.
+
+        Returns a tuple (line, did_timeout), where *did_timeout* is True
+        if the read timed out, and False otherwise.
         """
+        # Calls a private member because this is a different function based on
+        # the OS
         return self._readWithTimeout(f, timeout)
 
     def processOutputLine(self, line):
-        """Called for each line of output that a process sends to stdout/stderr.
-        """
+        """Called for each line of output that a process sends to stdout/stderr."""
         for handler in self.processOutputLineHandlers:
             handler(line)
 
@@ -693,15 +716,21 @@ falling back to not using job objects for managing child processes"""
             elif outputTimeout:
                 lineReadTimeout = outputTimeout
 
-            (line, self.didTimeout) = self.readWithTimeout(logsource, lineReadTimeout)
-            while line != "" and not self.didTimeout:
-                self.processOutputLine(line.rstrip())
+            (lines, self.didTimeout) = self.readWithTimeout(logsource, lineReadTimeout)
+            while lines != "":
+                for line in lines.splitlines():
+                    self.processOutputLine(line.rstrip())
+
+                if self.didTimeout:
+                    break
+
                 if timeout:
                     lineReadTimeout = timeout - (datetime.now() - self.startTime).seconds
-                (line, self.didTimeout) = self.readWithTimeout(logsource, lineReadTimeout)
+                (lines, self.didTimeout) = self.readWithTimeout(logsource, lineReadTimeout)
 
             if self.didTimeout:
-                self.proc.kill()
+                if self._kill_on_timeout:
+                    self.proc.kill()
                 self.onTimeout()
             else:
                 self.onFinish()
@@ -717,12 +746,16 @@ falling back to not using job objects for managing child processes"""
 
     def wait(self, timeout=None):
         """
-        Waits until all output has been read and the process is 
+        Waits until all output has been read and the process is
         terminated.
 
         If timeout is not None, will return after timeout seconds.
         This timeout only causes the wait function to return and
         does not kill the process.
+
+        Returns the process' exit code. A None value indicates the
+        process hasn't terminated yet. A negative value -N indicates
+        the process was killed by signal N (Unix only).
         """
         if self.outThread:
             # Thread.join() blocks the main thread until outThread is finished
@@ -732,7 +765,7 @@ falling back to not using job objects for managing child processes"""
                 self.outThread.join(timeout=1)
                 count += 1
                 if timeout and count > timeout:
-                    return
+                    return None
 
         return self.proc.wait()
 
@@ -745,7 +778,7 @@ falling back to not using job objects for managing child processes"""
 
     ### Private methods from here on down. Thar be dragons.
 
-    if mozinfo.isWin:
+    if isWin:
         # Windows Specific private functions are defined in this block
         PeekNamedPipe = ctypes.windll.kernel32.PeekNamedPipe
         GetLastError = ctypes.windll.kernel32.GetLastError
@@ -774,15 +807,33 @@ falling back to not using job objects for managing child processes"""
     else:
         # Generic
         def _readWithTimeout(self, f, timeout):
-            try:
-                (r, w, e) = select.select([f], [], [], timeout)
-            except:
-                # return a blank line
-                return ('', True)
+            while True:
+                try:
+                    (r, w, e) = select.select([f], [], [], timeout)
+                except:
+                    # return a blank line
+                    return ('', True)
 
-            if len(r) == 0:
-                return ('', True)
-            return (f.readline(), False)
+                if len(r) == 0:
+                    return ('', True)
+
+                output = os.read(f.fileno(), 4096)
+                if not output:
+                    output = self.read_buffer
+                    self.read_buffer = ''
+                    return (output, False)
+                self.read_buffer += output
+                if '\n' not in self.read_buffer:
+                    time.sleep(0.01)
+                    continue
+                tmp = self.read_buffer.split('\n')
+                lines, self.read_buffer = tmp[:-1], tmp[-1]
+                real_lines = [x for x in lines if x != '']
+                if not real_lines:
+                    time.sleep(0.01)
+                    continue
+                break
+            return ('\n'.join(lines), False)
 
     @property
     def pid(self):
@@ -824,16 +875,22 @@ class LogOutput(object):
 ### front end class with the default handlers
 
 class ProcessHandler(ProcessHandlerMixin):
+    """
+    Convenience class for handling processes with default output handlers.
+
+    If no processOutputLine keyword argument is specified, write all
+    output to stdout.  Otherwise, the function specified by this argument
+    will be called for each line of output; the output will not be written
+    to stdout automatically.
+
+    If storeOutput==True, the output produced by the process will be saved
+    as self.output.
+
+    If logfile is not None, the output produced by the process will be
+    appended to the given file.
+    """
 
     def __init__(self, cmd, logfile=None, storeOutput=True, **kwargs):
-        """
-        If storeOutput=True, the output produced by the process will be saved
-        as self.output.
-
-        If logfile is not None, the output produced by the process will be
-        appended to the given file.
-        """
-
         kwargs.setdefault('processOutputLine', [])
 
         # Print to standard output only if no outputline provided
